requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
nltk==3.8.1
textblob==0.17.1
schedule==1.2.0
flask==3.0.0
gunicorn==21.2.0
